requirements.txt
amqp==5.0.6
asgiref==3.3.4
autopep8==1.4.4
billiard==3.6.4.0
celery==5.1.1
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dateparser==0.7.2
Django==2.2.13
django-audit-log==0.7.0
django-crispy-forms==1.12.0
django-debug-toolbar==2.1
django-discover-runner==1.0
django-registration-redux==2.9
django-role-permissions==2.2.1
djangorestframework==3.12.4
fake-factory==9999.9.9
Faker==2.0.4
Flask==2.0.1
idna==2.8
itsdangerous==2.0.1
jet-bridge-base==0.7.7
jet-django==1.0.2
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
pep8==1.7.1
Pillow==6.2.1
prompt-toolkit==2.0.9
pycodestyle==2.5.0
pyflakes==2.3.1
python-dateutil==2.8.1
python-memcached==1.59
pytz==2018.5
regex==2019.11.1
requests==2.22.0
sentry-sdk==1.1.0
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
text-unidecode==1.3
tzlocal==2.0.0
urllib3==1.25.7
vine==5.0.0
wcwidth==0.1.7
Werkzeug==2.0.1
amqp==5.0.6
asgiref==3.3.4
astroid==2.4.1
autopep8==1.4.4
billiard==3.6.4.0
celery==5.1.1
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
dateparser==0.7.2
Django==3.2.4
django-audit-log==0.7.0
django-ckeditor==5.8.0
django-crispy-forms==1.12.0
django-debug-toolbar==2.1
django-discover-runner==1.0
django-filter==2.4.0
django-jet==1.0.8
django-js-asset==1.2.2
django-registration-redux==2.9
django-role-permissions==2.2.1
djangorestframework==3.12.4
fake-factory==9999.9.9
Faker==2.0.4
Flask==2.0.1
idna==2.8
isort==4.3.21
itsdangerous==2.0.1
jet-bridge-base==0.7.7
jet-django==1.0.2
Jinja2==3.0.1
kombu==5.1.0
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
pep8==1.7.1
Pillow==7.2.0
prompt-toolkit==2.0.9
pycodestyle==2.5.0
pyflakes==2.3.1
pylint==2.5.2
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.1
regex==2019.11.1
requests==2.22.0
sentry-sdk==1.1.0
six==1.14.0
SQLAlchemy==1.3.11
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
tzlocal==2.0.0
urllib3==1.25.7
vine==5.0.0
wcwidth==0.1.7
Werkzeug==2.0.1
wrapt==1.12.1
